Add existing files
This is version 2, which accounts for multiple instances of the application. This version allows multiple users to have their own local internal environment variables using the gradio.State objects. Users are able to run tasks that yield, with the help of the asyncio library. Next steps could include animating swaps.